ledger-saas/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

sqlalchemy==2.0.34
pydantic==2.8.2

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

pdfplumber==0.11.4
PyMuPDF==1.24.0
pytesseract==0.3.10
Pillow==10.4.0
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
